refactor: extract shared types, filter internal keys, unify formatting
